pylintrc: re-enable too-many-lines, part 6
To get under the 1000 line limit, move openstack
agent test classes into a different file.
